Fixed some compiler warnings Changed mod file Fixed the calculation of the sinus table for the balls

+++ src/demo/star_bmp.h
@@ -7,8 +7,8 @@ class StarBmp
 	//Initializes variables
 	StarBmp();
 
-	double x;
-	double y;
+	int x;
+	int y;
 	int spd;
-	double bmp;
+	int bmp;
 };
